Move post install logic to win32.lib
